Moved `ScalarType.BoolType` to OTS interfaces folder
